requirements.txt
Flask
gunicorn
mysql-connector-python
bcrypt
flask-cors
requests
